requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dacite==1.8.0
fastapi==0.95.0
graphql-core==3.2.3
h11==0.14.0
idna==3.4
pika==1.3.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strawberry-graphql==0.168.1
style==1.1.0
typing_extensions==4.5.0
update==0.0.1
urllib3==1.26.15
uvicorn==0.21.1
